Test de cliente controller al 97
He modificado la funcion de actualizar cliente ya que a la hora de cambiar el nombre del cliente la foto se guardaba con el nombre del cliente antiguo

+++ app/Http/Controllers/ClienteController.php
@@ -143,18 +143,17 @@ class ClienteController extends Controller
             'password' => 'nullable|string|min:8',
         ]);
 
-
         if ($request->hasFile('avatar')) {
             $image = $request->file('avatar');
-            $customName = 'perfil_' . $cliente->name . '.' . $image->getClientOriginalExtension();
-            if ($cliente->avatar) {
-                Storage::disk('public')->delete('images/' .$cliente->avatar);
+            $customName = 'perfil_' . ($validatedData['name'] ?? $cliente->name) . '.' . $image->getClientOriginalExtension();
+
+            if ($cliente->avatar && $cliente->avatar !== $customName) {
+                Storage::disk('public')->delete('images/' . $cliente->avatar);
             }
+
             $image->storeAs('images', $customName, 'public');
 
             $cliente->avatar = $customName;
-            $cliente->save();
-
         }
 
         $user = User::find($cliente->user_id);
@@ -166,12 +165,14 @@ class ClienteController extends Controller
             ]);
         }
 
-        Cache::forget($clienteCacheKey);
+        $cliente->save();
 
-        Cache::put($clienteCacheKey, $cliente, 20);
+        Cache::forget($clienteCacheKey);
+        Cache::put($clienteCacheKey, $cliente->fresh(), 20);
 
         return redirect()->route('clientes.index')->with('success', 'Cliente actualizado con éxito');
     }
+
 
 
     public function destroy($id)
@@ -194,10 +195,8 @@ class ClienteController extends Controller
             $user = User::find($cliente->user_id);
         }
 
-        // Eliminar el cliente
         $cliente->delete();
 
-        // Si el usuario existe, proceder con su eliminación
         if ($user) {
             $user->delete();
         }
